Import of fix to cache invalidation bug of tie::std::file.
Import of Michael Schlenker's application of Krzysztof's and
Tomasz's fix to prioqueue sorting.

Import of my changes to snit (Integration of a backport to
Tcl 8.3, by Kenneth Green, and correction of string handling
of list data.)

Import of Michael's changes to logger documentation and
examples.

Import of fix in docstrip testsuite.

Import of Pat's changes to des (Integration of Mac Cody's
TclDES(jr)).

Import of changes to list of installed applications.

Fixed testsuite problems of math::calculaus,
math::interpolate, and math::optimize by declaring the a
dependency on 8.4 due to use of 8.4isms. In code, package
index, testsuite, and documentation.