The ldapx module has a major problem in the "change" entries: they cannot represent all cases, and in particular those involving simultaneously a modification in attributes and a renaming of an entry.
For example:

    package require ldapx

    ::ldapx::entry create e
    ::ldapx::entry create c

    e dn uid=bob,o=org
    e set1 a v1
    e backup

    e dn uid=sam,o=org
    e add1 a v2

    e debug
    # =>
    #     dn = <uid=sam,o=org>
    #     format = standard
    #             a: v1 v2
    #     backup dn = uid=bob,o=org
    #             a: v1
    # There is both a renaming of the entry (bob->sam) and a new
    # attribute value (v2) for attribute a

    c diff e
    puts [c change]
    # =>
    #     add {a {v1 v2}}
    # The renaming information is lost!

If you try again without the line "e add1 a v2", you can see that
the change entry is correct, as it contains a "modrdn" meaning the
renaming of the entry.

So, the format of the "change" entry is incompatible with the
described functionnality. This patch modify the format of the
"change" entry from a single action to a list of actions.  In the
above example, it returns be:
	{ {add {a {v1 v2}}} {modrdn ... } }
instead of :
	{add {a {v1 v2}}}

This format modification should not a real problem, since "change"
entries are not intended to be used by programs, they are designed
only for used by ldif and ldap classes.

The ldapx version number is upgraded from 0.2.5 to 1.0 to reflect
the potential incompatibility.