The packages fileutil::decode and zipfile::*, while residing in the AS/perforce had versions 0.1 and 0.2, with the change number attached later on installation, making them actually version 0.1.nnn, etc.
During the move of the sources to Tcllib the numbers should have been
incremented to properly distinguish the code in the new location from
these old revisions. This did not happen, making the tcllib variant
look like older/lesser revisions.

This change fixes the issue now and bumps the version numbers, giving
the Tcllib sources precedence.