Disabled the i_error_nonterminal instruction in all implementations of the RDE runtime (Tcl snit|oo, Critcl).
This means that at the moment the error state of a pt-made parser will
only contain messages with references to expected terminal symbols.
Non-terminal symbols will not occur. This may change in the future,
when it is worked out how to handle them.

Updated the testsuite, as less as there is, to match this.