Added a procedure for normalizing the name of a class to it's globally addressible name when interacting with oo::meta::* namespace commands
Introducing the concept that dict keys which terminate in a : are a sign that they are
leaves, and no dicts will be nexted beneath them.

Dict merge turned out to be entirely too naive a system for combining properties
across nested dicts. Using the name and possible failures in [dict size], the system
now drills down several levels of nested dicts to ensure at least the "option" and "const"
subdict merge correctly.

The "property" keyword has been modifed. If given 2 arguments, it files a property under
the "const" section instead of under the root. The property keyword ALSO marks all fields
created by it as leaves.

Added an "option" keyword to classes. The keyword simply files new entries in the "option"
meta section.

Added a "meta cget" method which will search for a property in the local config dict,
the const dict of its meta data, or the root of its meta data, mapping the field to field:
if necissary.

Updated all of the oo::meta and oo::options tests to reflect this new behavior.