Changed the options array to a config dict. There's a ton of code on Tao and IRM that rely on data being stored in dict format. (Sorry Will!)
Added a "config" ensemble to manage configuration for the object.

"configlist" replaced with "config merge"
"configlist_triggers" replaced with "config triggers"

"configure" not does rougly the same thing as "config set"
"cget" is now an alias for "config get"

Added a tool_datatype to manage the trandformations of basic integers, reals, vectors,
and common time stamps.

Stripped out the more esoteric pipeline concepts from TOOL. Those features are now 
in TAO, and will mark the transition point between the two frameworks. (TOOL is intended
to be the 20% solution that gets 80% done.)

Added option_class handling to TOOL

Added tests for all of the major features thusfar.