Fixes to http based on testing with Internet Explorer running on localhost:
The httpd module now uses non-blocking IO where possible, especially when establishing
an initial connection. 

Replaced calls to "puts" with "chan puts", to prevent errors in console apps where the 
"::puts" command is redirected.

The output method now waits for the channel to be writeable before trying to send the reply.