Updated practcl to 0.3
Added ::practcl::os, to autodetect basic operating system information in the absence of autoconf. Needed for odie (and projects base on kitbuilders) where the local autoconf relies on Tcl, and Tcl hasn't been compiled yet, and practcl is doing the downloading and compiling of Tcl Added an "smake" like target/dependency system. Users can build target objects, and give them dependencies and triggers. The actual code executes to resolved the dependencies and triggers is not implemented explicetly, instead the use is given a ::make array with a list of potentail targets with a true/false if the need to be built in this pass. The new target_obj class shares some core functions with ::practcl::object, but they are not full-fledged practcl objects. Added a mechanism for a static executable to build a table of contents for static packages that are included. TK is now build statically and linked into the final executable. Added a fix for tclconfig to prevent it from trying to unpack itself inside of itself Added a config.sh method for core libraries. This dumps out the XXXConfig.sh built by autoconf into a key/value list for easier parsing.