httpd modifications:
The manual file for httpd is now generated dynamically during the call to build/build.tcl and
is developed by scanning the source files and scraping comments.

The test.tcl file generated during httpd.test is now deleted prior to ending the test.

Redistributed the hard coded documentation files as comments in front of the methods
they describe to be assembled by the manual file generator.