Added the ability for clay, httpd, and practcl to auto-generate the HTML form of their documentation and slip that file into the appropriate place in tcllib to present live embedded docs.
Documentation fixes

Removed stale tests from tool

Moved the auto-documenter function out of clay and into practcl. If your 
are going to have a build helper around, why not put it there? Build scripts in
tcllib will look for a resident copy of the source and failing that, ask
for a version of practcl that the function is known to reside it.