ncgi - Tkt [1f900bdf6b] - B, T Bumped to version 1.4.4. Regenerated docs.
Applied patch to improvde handling of 1-byte encodings.
This is the second of two.
Thanks to https://saschaszott.github.io/
Updated test case ncgi-3.10.

His explanation:

In practice www-url-encoded POST params can use encodings other than
UTF-8 (think of legacy Tcl applications that use one of the ISO-8859-x
charsets). In this case URL parameters can contain references to 8-bit
code points (in the form of `%[A-F0-9][A-F0-9]`) that are not valid
UTF-8 code points.

For example, `%DC` can be used as a percent encoding for the german umlaut `Ü`
(if a Tcl application is based on ISO-8859-1). Currently, the `decode`
procedure does not decode `%DC` as all one byte UTF-8 code points must start
with `[0-7]`.

This commit improves the handling of one byte percent encoded non-ASCII
characters. It allows to use ncgi in application contexts that do not use
UTF-8 as the base encoding.