struct - struct::record - Tkt [baa334cfef], B, I, T Version 1.2.2
Fixed broken deletion of record instances. The old code tried to
perform a quick deletion by command list, and somewhere it went of the
rails so that a nested instance lost the alias yet kept the field
values around.

The new code does it by the numbers, the reverse of creation, by
iterating over the list of member and unsetting or recursively
deleting each, per their definition.

Additional cleanup using `upvar 0` to import state from record
namespaces, reducing quoting hell.

Further fixed the semi-broken implementation of `exists`. The first
condition in each branch was completely broken and it somewhat worked
only because of the second condition checking for (alias) commands. I
suspect that this was also the cause of the issue with deletion.

Replaced with much simpler code checking the relevant package state
variables where this information is directly accessible. The broken
conditions sort of tried that, except they missed importing the proper
variable into the scope, looked at the wrong variable, and got
confused in the quoting hell.

Version bump to 1.2.2 (B).