Continued work on release 1.20.
Ticket [138d56dbaf].

Extended module aggregator with code to detect a tcllibc and register
it it as installable package.

Extended set of install actions with one placing the entire module
into the installation directory.

Fixed localdoc code to supply the `distribution` variable the new
detection code requires to exist.

Extended the installation guide with the note that building tcllibc
now also causes it to be installed.

Bumped the bundle version to 1.20.