dns - ipMore - <I> Bumped ip to 1.5, ipMorec to 1.1.
Removed the code skipping Windows/SunOS 5.6. It is not explained why
it is needed, and why only for the specific platforms. This means that
the accelerator code does not contain a compile-time build system
choice any longer.

If the disabling is truly required then compile-detection can be added
back, plus any necessary signaling to the Tcl side and runtime.

Further, moved the command definitions out of the namespace eval, and
added the namespacing back in through the command names itself.
Reduces indentation and looks generally nicer.

Lastly, fixed the issues in the test suite which prevented it from
properly testing the C parts even when available.