httpd - B - Ticket [4efe78a285] Handle fumagic filetype result differences.
Bumped to version 4.3.5. Still fallout from the fumagic filetype
restructuring and result structure revamp. Modified httpd to handle
both forms of results, based on the fumagic version.

Thanks to submitter @fpigorsch for the report and patch. (Even if I
did not use that patch and went for a more general fix).