markdown - B, EF - Ticket [29a9996f1e] Sync Tclssg markdown over to Tcllib, as feasible.
Added the Tclssg markdown testsuite to the module.

Extended the handling of fenced code blocks. Now supports extended
converters which see the arguments after th lang specififer in the
code blocks info string.

Added tests for the converter handling.

Bumped to version 1.2

Not everything could be synced over. The two code bases have
definitely diverged in their handling of fenced code blocks.
Each is keeping to their own on this now.

Thanks to submitter @dbohdan for the report and references.