coroutine - B, EF - Tickets [bf8b80af3d], and [adf34e9a7e]
Fixed coroutine utility command `global` for when it is invoked from
the #1 frame (== main procedure of the coroutine).

Accepted patch adding utilities and auto-wrappers for `puts` and
`socket`. Wrote documentation for them.

Version bumps:
  - coroutine       to 1.3
  - coroutine::auto to 1.2

Cleanup of trailing whitespace.
Regenerated embedded documentation.

Thanks to submitter <shawnw.mobile@gmail.com> for the reports and
patches.