cmdline - B - Fix test breakage due to cleanup of trailing whitespace.
Decided to fix the code instead of having the test rely on trailing
whitespace (*). Does not generate trailing whitespace in the usage
texts anymore.

Bumped version to 1.5.1.
Regenerated documentation.

(*) It is bad enough that Markdown uses trailing whitespace to encode
hard line breaks, forcing things for that doctools output format.