textutil - string - I - fixed indentation of benchmark code
textutil - wcswidth - I - replaced core type/char functions with new implementations roughly 30x faster.

The new implementation uses a nested binary tree of if commands to quickly bisect the result
domain. Interestingly enough this is faster than using a 2-stage table lookup. And while it is not
as fast as a single-stage table lookup (~45x) it also does not come with the horrendous size
overhead of the same (~120x).

The research leading to this (*) is saved in sub directory `build/alternates`. See
`build/alternates/README.md` for the summary.

(*) Scripts, results (.csv(.xz)), GNUplot configurations, ...